move split code to another file
